deleting from category

+++ main.py
@@ -9,32 +9,6 @@ import click
 def mycommands():
     pass
 
-
-# @mycommandss.group(name='inventory', help="Inventory Management Commands")
-# def inventory_group():
-#     pass
-#
-#
-# @mycommandss.group(name='customer', help="Customer Management Commands")
-# def customer_group():
-#     pass
-#
-#
-# @mycommandss.group(name='supplier', help="Supplier Management Commands")
-# def supplier_group():
-#     pass
-#
-#
-# @mycommandss.group(name='product', help="Product Management Commands")
-# def product_group():
-#     pass
-
-
-categories = session.query(Category).all()
-suppliers = session.query(Supplier).all()
-products = session.query(Product).all()
-# _------------------------------ Create Commands ------------------------
-
 '''----------------------- C A T E G O R I E S ______________________--'''
 
 
@@ -46,15 +20,16 @@ def view_categories():
 @mycommands.command()
 @click.option('--name', '-n', prompt="Enter the category name")
 def new_category(name):
+    categories = session.query(Category)
     if name not in categories:
         new_category = Category(
             name=name
         )
         session.add(new_category)
         session.commit()
-        click.echo(f"Category '{name}' successfully added!")
+        click.echo(click.style(f"--------------- CATEGORY SUCCESSFULLY ADDED ----------------", fg='green', bold=True))
     else:
-        click.echo(f"Category '{name}' already exists.")
+        click.echo(click.style(f"*** Category '{name}' already exists. ***", fg='red'))
 
 
 @mycommands.command()
@@ -68,22 +43,64 @@ def update_category(search_name, new_name):
             Category.name: f'C-{new_name}'
         })
         session.commit()
-        click.echo("Name successfully updated!")
+        click.echo(click.style("-------------- NAME SUCCESSFULLY UPDATED -------------", fg='green',bold=True ))
     else:
-        click.echo("Sorry! That category does not exist and cannot be updated.")
+        click.echo(click.style(" *** Sorry! That category does not exist and cannot be updated. ***", fg='red', bold=True))
 
 
 @mycommands.command()
 @click.option('--search_name', '-sn', prompt="Enter the category name to delete")
 def delete_category(search_name):
     category_to_delete = session.query(Category).filter(Category.name.like(f'%{search_name}%')).first()
-    click.echo(f"{category_to_delete}")
     if category_to_delete:
-        session.delete(category_to_delete)
-        session.commit()
-        click.echo("Category successfully deleted!")
+        confirmation = click.prompt(
+            "To complete delete, related product records will be updated to a different supplier via auto-assign or selected supplier. Continue? Y/N")
+        if confirmation.lower() == 'y':
+            category_update_mode = click.prompt(
+                "Update via auto-assign or selected supplier? Enter 1 or 2 to choose.\n1. Auto-assign (1)\n2. Select Category(2)\nEnter Selection",
+                type=int)
+            if category_update_mode == 1:
+                products_to_update_category = session.query(Product).filter(
+                    Product.category_id == category_to_delete.id).all()
+                if products_to_update_category:
+                    categories = session.query(Supplier).filter(Supplier.id != category_to_delete.id).all()
+                    random_category = random.choice(categories)
+                    session.query(Product).filter(Product.category_id == category_to_delete.id).update(
+                        {
+                            Product.category_id: random_category.id
+                        }
+                    )
+                    click.echo(click.style(f"Product categories successfully updated to {random_category}", fg='green',
+                                           bold=True))
+                    session.delete(category_to_delete)
+                    session.commit()
+                    click.echo(click.style("------------------- CATEGORY SUCCESSFULLY DELETED ---------------------",
+                                           fg='green', bold=True))
+                else:
+                    click.echo(click.style("No related products found!",fg='red', bold=True))
+            elif category_update_mode == 2:
+                new_category_name = click.prompt(click.style("Which category would you like to transfer the product records to?",fg='cyan', bold=True))
+                new_category = session.query(Category).filter(Category.name.like(f'%{new_category_name}%')).first()
+                session.query(Product).filter(Product.supplier_id == category_to_delete.id).update(
+                    {
+                        Product.category_id: new_category.id
+                    }
+                )
+                click.echo(
+                    click.style(f"Product categories successfully updated to {new_category}!", fg='green', bold=True))
+                session.delete(category_to_delete)
+                session.commit()
+                click.echo(
+                    click.style("------------------- CATEGORY SUCCESSFULLY DELETED ---------------------", fg='green',
+                                bold=True))
+            else:
+                click.command(click.style("ERROR! Input Invalid", fg='red', bold=True))
+        elif confirmation.lower() == 'n':
+            click.echo(click.style("Delete Action Aborted!", fg='red', bold=True))
+        else:
+            click.echo(click.style("ERROR! Invalid input!", fg='red', bold=True))
     else:
-        click.echo("Sorry! That category does not exist and cannot be deleted.")
+        click.echo(click.style("ERROR! That category does not exist and cannot be deleted.", fg='red', bold=True))
 
 
 @mycommands.command()
@@ -104,13 +121,14 @@ def view_suppliers():
 @mycommands.command()
 @click.option('--name', '-n', prompt="Enter the supplier name")
 def new_supplier(name):
+    suppliers = session.query(Supplier)
     if name.lower() not in [supplier.name.lower() for supplier in suppliers]:
         new_supplier = Supplier(
             name=name
         )
         session.add(new_supplier)
         session.commit()
-        click.echo(f"Supplier '{name}' successfully added!")
+        click.echo(click.style(f"---------------- SUPPLIER '{name}' SUCCESSFULLY ADDED ---------------", fg='green', bold=True))
     else:
         click.echo(f"Supplier '{name}' already exists.")
 
@@ -138,7 +156,7 @@ def delete_supplier(search_name):
     if supplier_to_delete:
         confirmation = click.prompt("To complete delete, related product records will be updated to a different supplier via auto-assign or selected supplier. Continue? Y/N")
         if confirmation.lower() == 'y':
-            supplier_update_mode = click.prompt("Update via auto-assign or selected supplier? Enter 1 or 2 to choose.\n1. Auto-assign (1)\n2. Select Supplier(2)\nEnter Selection:", type=int)
+            supplier_update_mode = click.prompt("Update via auto-assign or selected supplier? Enter 1 or 2 to choose.\n1. Auto-assign (1)\n2. Select Supplier(2)\nEnter Selection", type=int)
             if supplier_update_mode == 1:
                 products_to_update_supplier = session.query(Product).filter(Product.supplier_id == supplier_to_delete.id).all()
                 if products_to_update_supplier:
@@ -149,10 +167,11 @@ def delete_supplier(search_name):
                             Product.supplier_id: random_supplier.id
                         }
                     )
-                    click.echo(f"Product Suppliers successfully updated to {random_supplier}!")
+                    click.echo(click.style(f"Product Suppliers successfully updated to {random_supplier}!", fg='green',bold=True))
                     session.delete(supplier_to_delete)
                     session.commit()
-                    click.echo("------------------- SUPPLIER SUCCESSFULLY DELETED ---------------------")
+                    click.echo(click.style("------------------- SUPPLIER SUCCESSFULLY DELETED ---------------------",
+                                           fg='green', bold=True))
                 else:
                     click.echo("No related products found!")
             elif supplier_update_mode == 2:
@@ -163,18 +182,18 @@ def delete_supplier(search_name):
                         Product.supplier_id: new_supplier.id
                     }
                 )
-                click.echo("Product Suppliers successfully updated!")
+                click.echo(click.style(f"Product Suppliers successfully updated to {new_supplier}!", fg='red', bold=True))
                 session.delete(supplier_to_delete)
                 session.commit()
-                click.echo("------------------- SUPPLIER SUCCESSFULLY DELETED ---------------------")
+                click.echo(click.style("------------------- SUPPLIER SUCCESSFULLY DELETED ---------------------", fg='green', bold=True))
             else:
-                click.command("ERROR! Input Invalid")
+                click.command(click.style("ERROR! Input Invalid", fg='red', bold=True))
         elif confirmation.lower() == 'n':
-            click.echo("Delete Action Aborted!")
+            click.echo(click.style("--------------- Delete Action Aborted ----------------",fg='red', bold=True))
         else:
-            click.echo("ERROR! Invalid input!")
+            click.echo(click.style("ERROR! Invalid input!", fg='red', bold=True))
     else:
-        click.echo("Sorry! That supplier does not exist and cannot be deleted.")
+        click.echo(click.style("Sorry! That supplier does not exist and cannot be deleted.", fg='red', bold=True))
 
 
 @mycommands.command()
@@ -200,7 +219,7 @@ def view_product_details(name):
             click.echo(
                 f'({product.id}): Name:{product.name}, Price:{product.price} | Quantity:{product.quantity} | Category:{product.category.name} | Product:{product.supplier.name}')
         else:
-            click.echo("Product you searched fo does not exist.")
+            click.echo(click.style("Product you searched for does not exist.", fg='red', bold=True))
 
 
 @mycommands.command()
@@ -224,9 +243,9 @@ def add_product(name, category, supplier, price, quantity):
         )
         session.add(new_product)
         session.commit()
-        click.echo("New product added successfully!")
+        click.echo(click.style("------------- NEW PRODUCT ADDED SUCCESSFULLY -------------!", fg='green', bold=True))
     else:
-        click.echo("Sorry! Category or Supplier you entered does not exist.")
+        click.echo(click.style("Sorry! Category or Supplier you entered does not exist.",fg='green', bold=True))
 
 
 @mycommands.command()
@@ -236,14 +255,12 @@ def delete_product(name):
     if product_to_delete:
         session.delete(product_to_delete)
         session.commit()
-        click.echo("Product has been successfully deleted")
+        click.echo(click.style("------------- PRODUCT HAS BEEN SUCCESSFULLY DELETED ------------", fg='green', bold=True))
     else:
-        click.echo("Error! No such product exists.")
+        click.echo(click.style("Error! No such product exists.", fg='green', bold=True))
 
 
 product_details = ("name", "price", "quantity", "category", "supplier")
-
-
 
 @mycommands.command()
 @click.option('--choice', '-n', prompt="What would you like to update? Select", type=click.Choice(product_details))
@@ -271,7 +288,7 @@ def update_product(choice, name):
                 update_data['supplier_id'] = supplier_record.id
             else:
                 click.echo(
-                    "Entered supplier does not exist. View existing products using the 'view-suppliers' command.")
+                    click.style("Entered supplier does not exist. View existing products using the 'view-suppliers' command.", fg='red', bold=True))
         if choice.lower() == 'category':
             update_to_category = click.prompt("Enter the new category")
             category_record = session.query(Category).filter(Category.name.like(f'%{update_to_category}%')).first()
@@ -279,11 +296,11 @@ def update_product(choice, name):
                 update_data['category_id'] = category_record.id
             else:
                 click.echo(
-                    "Entered category does not exist. View existing categories using the 'view-categories' command.")
+                    click.style("Entered category does not exist. View existing categories using the 'view-categories' command.", fg='red', bold=True))
         print(update_data)
         session.query(Product).filter_by(id=product_to_update.id).update(update_data)
         session.commit()
-        click.echo("Product is successfully updated!")
+        click.echo(click.style("----------- PRODUCT IS SUCCESSFULLY UPDATED --------------", bg='green', bold=True))
     else:
         click.echo(
             "Sorry! That product does not exist and cannot be updated. View existing products using the 'view-product-details' command.")
@@ -328,13 +345,12 @@ def add_purchase(customer_name, product_name, quantity):
             })
             session.commit()
         else:
-            click.echo(f"-------------- ERROR ------------\n"
-                       f"Unable to purchase. Only {product.quantity} item remaining to be purchased!")
+            click.echo(click.style(f"ERROR! Unable to purchase. Only {product.quantity} item remaining to be purchased!", fg='red',bold=True))
 
         session.commit()
-        click.echo("----------- PURCHASE SUCCESSFULLY ADDED ------------")
+        click.echo(click.style("----------- PURCHASE SUCCESSFULLY ADDED ------------", fg='green', bold=True))
     else:
-        click.echo("ERROR! Product or Customer you entered does not exist")
+        click.echo(click.style("ERROR! Product or Customer you entered does not exist", fg='red', bold=True))
 
 
 @mycommands.command()
@@ -345,7 +361,7 @@ def view_customer_purchase_details(name):
         for purchase in customer.purchases:
             click.echo(f'({purchase.id}) Customer: {purchase.customer.full_name} | Product: {purchase.product.name} | Qty: {purchase.quantity}\n')
     else:
-        click.echo("ERROR! Entered customer was not found.")
+        click.echo(click.style("ERROR! Entered customer was not found.", fg='red', bold=True))
 
 
 @mycommands.command()
@@ -356,7 +372,7 @@ def view_product_purchase_details(name):
         for purchase in product.purchases:
             click.echo(f'({purchase.id}) Customer: {purchase.customer.full_name} | Product: {purchase.product.name} | Qty: {purchase.quantity}\n')
     else:
-        click.echo("ERROR! Entered product was not found.")
+        click.echo(click.style("ERROR! Entered product was not found.", fg='green', bold=True))
 
 
 if __name__ == '__main__':
